Phrase broken out to separate file.
